ch 14.2

+++ Sales_data.h
@@ -66,7 +66,7 @@ private :
 
 	std::string bookNo;
 	unsigned units_sold = 0;
-	unsigned price = 0 ;
+	double price = 0 ;
 	double revenue = 0.0;
 };
